requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
pandas==2.2.3
openpyxl==3.1.5
pyarrow==19.0.1
duckdb==1.2.1
PyYAML==6.0.2
python-dotenv==1.0.1
kiteconnect==5.0.1
httpx==0.28.1
plotly==6.0.1
python-multipart==0.0.20
kiteconnect
streamlit
openpyxl